+++ source/c/portable.h
@@ -20,13 +20,13 @@
 /* define integer type names for portability */
 /* use 'sint' and 'uint' unless space is at a premium */
 
-typedef long int		sint;
+typedef int		        sint;
 typedef char			sint8;
 typedef short int		sint16;
-typedef long int		sint32;
+typedef int		        sint32;
 typedef unsigned char		uint8;
 typedef unsigned short int	uint16;
-typedef unsigned long int	uint32;
+typedef unsigned int	        uint32;
 typedef unsigned char		bool;
 
 /* tiny uses arrays of characters to buffer I/O at various points.
